+++ karma.conf.js
@@ -59,7 +59,7 @@ module.exports = function(config) {
             'spec/wavesurfer.spec.js',
             'spec/peakcache.spec.js',
             'spec/mediaelement.spec.js',
-            'spec/webaudio-media.spec.js'
+            'spec/mediaelement-webaudio.spec.js'
         ],
         customHeaders: [
             {
@@ -74,7 +74,7 @@ module.exports = function(config) {
             'spec/wavesurfer.spec.js': ['webpack'],
             'spec/peakcache.spec.js': ['webpack'],
             'spec/mediaelement.spec.js': ['webpack'],
-            'spec/webaudio-media.spec.js': ['webpack'],
+            'spec/mediaelement-webaudio.spec.js': ['webpack'],
 
             // source files, that you want to generate coverage for
             // do not include tests or libraries