+++ spec/mediaelement-webaudio.spec.js
@@ -3,6 +3,6 @@
 import { sharedTests } from './mediaelement-mediaelement_webaudio-shared';
 
 /** @test {WaveSurfer} */
-describe('WaveSurfer/MediaElement:', function() {
-    sharedTests('MediaElement');
+describe('WaveSurfer/MediaElementWebAudio:', function() {
+    sharedTests('MediaElementWebAudio');
 });