+++ spec/webaudio-media.spec.js
@@ -1,8 +1,0 @@
-/* eslint-env jasmine */
-
-import { sharedTests } from './mediaelement-webaudiomedia-shared';
-
-/** @test {WaveSurfer} */
-describe('WaveSurfer/WebAudioMedia:', function() {
-    sharedTests('WebAudioMedia');
-});